alter cache metadata files
