+++ src/main/resources/workflows/callCaching/A_cacheWithinWF/A_cacheWithinWF.metadata
@@ -1,0 +1,14 @@
+{
+  "cache": {
+    "workflowName": "cacheWithinWF",
+    "cacheWithinWF.four": {
+      "cache": {
+        "cacheHitCall": "cacheWithinWF.two"
+      }
+    }
+  },
+  "assertEqual": {
+    "cacheWithinWF.twoAgain.inputs": "cacheWithinWF.fourAgain.inputs",
+    "cacheWithinWF.twoAgain.outputs": "cacheWithinWF.fourAgain.outputs"
+  }
+}
