+++ src/main/resources/workflows/callCaching/B_cacheBetweenWF/B_cacheBetweenWF.metadata
@@ -3,25 +3,21 @@
     "workflowName": "cacheBetweenWF",
     "cacheBetweenWF.threeAgain": {
       "cache": {
-        "allowResultReuse": true,
         "cacheHitCall": "cacheWithinWF.three"
       }
     },
     "cacheBetweenWF.twoAgain": {
       "cache": {
-        "allowResultReuse": true,
         "cacheHitCall": "cacheWithinWF.two"
       }
     },
     "cacheBetweenWF.fourAgain": {
       "cache": {
-        "allowResultReuse": true,
         "cacheHitCall": "cacheWithinWF.two"
       }
     },
     "cacheBetweenWF.oneAgain": {
       "cache": {
-        "allowResultReuse": true,
         "cacheHitCall": "cacheWithinWF.one"
       }
     }
