+++ src/main/resources/workflows/callCaching/B_cacheBetweenWF/B_cacheBetweenWF.wdl
@@ -1,5 +1,5 @@
 task oneAgain {
-  Int radius = 51
+  Int radius = 62
     command {
         echo ${radius*radius}
     }
